Manually cast system tv_sec to i64
Fixes #18.

+++ src/system.rs
@@ -73,7 +73,7 @@ fn file_time_as_u64(ft: &FILETIME) -> u64 {
 pub unsafe fn sys_time() -> (i64, i16) {
     let mut tv = libc::timespec { tv_sec: 0, tv_nsec: 0 };
     let _ = clock_gettime(libc::CLOCK_REALTIME, &mut tv);
-    (tv.tv_sec, (tv.tv_nsec / 1000) as i16)
+    (tv.tv_sec as i64, (tv.tv_nsec / 1000) as i16)
 }
 
 /// Returns the system’s current time, as a tuple of seconds elapsed since
